% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_embeddings.R
\name{clean_embeds}
\alias{clean_embeds}
\title{Quickly reformat a sentence transformers embedding output}
\usage{
clean_embeds(embeds, ...)
}
\arguments{
\item{embeds}{an n-dimensional data frame with an index column}

\item{...}{columns that should be dropped, i.e. document or index column}
}
\value{
data frame with column names cleaned and index column removed
}
\description{
Quickly reformat a sentence transformers embedding output
}
