% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_entities.R
\name{clean_entities}
\alias{clean_entities}
\title{Quickly extract columns of interest from Hugging Face entity recognisers}
\usage{
clean_entities(df)
}
\arguments{
\item{df}{data frame with entities extracted in JSON format}
}
\value{
tidier data frame for analysis.
}
\description{
Quickly extract columns of interest from Hugging Face entity recognisers
}
