% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_grabber.R
\name{context_grabber}
\alias{context_grabber}
\title{Retrieve the 1-5 word boundaries before and 1-10 after a pattern.}
\usage{
context_grabber(
  data,
  text_var = mention_content,
  pattern,
  range_before = 5,
  range_after = 10,
  show_pattern = FALSE
)
}
\arguments{
\item{data}{Data frame or tibble object}

\item{text_var}{Name of the text variable}

\item{pattern}{Pattern to search for - should be a character vector}

\item{range_before}{Number of words before}

\item{range_after}{Number of words after}

\item{show_pattern}{Whether to print the pattern as well as the context tibble}
}
\value{
a tibble of the extracted contexts
}
\description{
Boundaries are broken up by spaces (super basic)
}
