% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_multiple.R
\name{count_multiple}
\alias{count_multiple}
\title{Count multiple variables of the same data type}
\usage{
count_multiple(df, ...)
}
\arguments{
\item{df}{Data frame or tibble object}

\item{...}{The variables you want to create individual tables for - unquoted and should be of the same data type.}
}
\value{
Summary frames which separately count inputted variables and add a percent column
}
\description{
Count multiple variables of the same data type
}
