% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gt_summary.R
\name{disp_gt_sent_time}
\alias{disp_gt_sent_time}
\title{Title}
\usage{
disp_gt_sent_time(
  df,
  sentiment_var = sentiment,
  date_var = date,
  chart_type = c("lines", "bars"),
  time_unit = c("week", "day", "month", "quarter", "year"),
  date_breaks = "4 months",
  date_labels = "\%b"
)
}
\arguments{
\item{date_labels}{}
}
\description{
Title
}
