% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gt_summary.R
\name{disp_gt_theme}
\alias{disp_gt_theme}
\title{quickly add re-usable theme elements for gt plot funcs}
\usage{
disp_gt_theme()
}
\value{
list of ggplot boiler plate theme options
}
\description{
quickly add re-usable theme elements for gt plot funcs
}
