% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gt_summary.R
\name{disp_gt_vot}
\alias{disp_gt_vot}
\title{Title}
\usage{
disp_gt_vot(
  df,
  date_var,
  time_unit = c("day", "week", "month", "quarter", "year"),
  bar_colour = "#628EFD",
  date_breaks = "4 months",
  date_labels = "\%b"
)
}
\arguments{
\item{date_labels}{}
}
\description{
Title
}
