% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_LDAs_lollipop.R
\name{explore_LDAs_lollipops}
\alias{explore_LDAs_lollipops}
\title{explore_LDAs but with lollipops!}
\usage{
explore_LDAs(
 ldas,
 top_terms = TRUE,
 top_n = 15,
 nrow = 2,
 diff_terms = TRUE,
 diff_n = 10,
 bigrams = TRUE,
 bigram_n = 50,
 bigram_threshold = 0.5,
 exemplars = TRUE,
 exemplars_n = 50,
 probabilities = TRUE,
 coherence = TRUE,
 coherence_n = 10
)
}
\arguments{
\item{ldas}{A nested tibble where each row contains and LDA model.}

\item{top_terms}{Should bar charts of the top terms for each topic be produced?}

\item{top_n}{How many terms should be included in the top terms bar chart for each topic?}

\item{nrow}{How many rows should the facetted plot go across?}

\item{diff_terms}{Should bar charts of distinguishing terms for each topic be produced?}

\item{diff_n}{How many terms should be included in the distinguishing terms bar chart for each topic?}

\item{bigrams}{Should a bigram network be produced for each topic?}

\item{bigram_n}{How many bigrams should be included in the network?}

\item{bigram_threshold}{What proportion of a post should be assigned to a topic for it to be considered for inclusion in the bigram network?}

\item{exemplars}{Should a tibble of exemplar posts be produced?}

\item{exemplars_n}{How many exemplars should be provided for each topic?}

\item{probabilities}{Should a tibble of document-topic probabilities be produced?}

\item{coherence}{Should a tibble of topic coherence measurements be produced?}

\item{coherence_n}{Number of words to be used in coherence calculation.}
}
\value{
A nested tibble containing columns with the requested objects.
}
\description{
Produces exploratory visualisations from LDA models.
}
