% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_radarly.R
\name{fix_radarly}
\alias{fix_radarly}
\title{Quickly wrangle a couple of annoying columns from Radarly outputs}
\usage{
fix_radarly(df)
}
\arguments{
\item{df}{Name of Data Frame / Tibble object}
}
\value{
Edited Data Frame
}
\description{
Quickly wrangle a couple of annoying columns from Radarly outputs
}
