% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_setup.R
\name{folder_setup}
\alias{folder_setup}
\title{Quickly make the reequisite folders in the data_science_project_work folder for a project}
\usage{
folder_setup(project_name, subdir = "microsoft/project_work")
}
\arguments{
\item{project_name}{Project's name, e.g. 'sustainbaility'}

\item{subdir}{The company + any additional extensions}
}
\value{
logicals telling you whether folders and files were created
}
\description{
Quickly make the reequisite folders in the data_science_project_work folder for a project
}
\examples{
\dontrun{
folder_setup("sustainability", subdir = "company_x/project_folders")
}
}
