% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Retrieve hashtags and at_tags from text}
\usage{
get_tags(data, text_var = mention_content)
}
\arguments{
\item{data}{Data frame or tibble object}

\item{text_var}{Name of the text variable}
}
\value{
A list of two tibbles
}
\description{
Returns a list of two tibbles - hashtags and at_tags.
Index into the list with either $ syntax or []
}
