% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huggingface_quick_clean.R
\name{huggingface_quick_clean}
\alias{huggingface_quick_clean}
\title{Run some fast text normalisation steps for output to Huggingface}
\usage{
huggingface_quick_clean(df, text_var)
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{text_var}{Name of the text variable}
}
\value{
Tibble with two columns
}
\description{
Run some fast text normalisation steps for output to Huggingface
}
