% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bigram_viz.R
\name{make_bigram_viz}
\alias{make_bigram_viz}
\title{Plot a bigram network}
\usage{
make_bigram_viz(data, text_var = mention_content, top_n = 50, min = 10, ...)
}
\arguments{
\item{data}{Data frame or tibble object}

\item{text_var}{Name of the text variable}

\item{top_n}{Number of n-grams to plot}

\item{min}{Min frequency of n-gram}
}
\value{
A ggplot2 network viz
}
\description{
Plot a bigram network
}
