% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DTMs_parallel.R
\name{make_DTMs_parallel}
\alias{make_DTMs_parallel}
\title{Run the \code{make_DTMs} function logic but with parallel processing!}
\usage{
make_DTMs_parallel(df, text_var, min_freq = 10, hashtags = FALSE, mentions = FALSE)
}
\arguments{
\item{df}{A data frame or tibble where each row is a separate post.}

\item{text_var}{The variable containing the text which you want to explore.}

\item{min_freq}{The minimum number of times a term must be observed to be considered.}

\item{hashtags}{Should hashtags be included?}

\item{mentions}{Should mentions be included?}
}
\value{
A nested tibble in which each row contains a document-term matrix.
}
\description{
Function does not implictly clean the text variable like \code{make_DTMs}
}
\examples{
make_DTMs_parallel(
 df = sprinklr_export,
 text_var = Message,
 min_freq = 25,
 hashtags = FALSE,
 mentions = FALSE
)
}
