% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gt_summary.R
\name{make_gt_grad_pal}
\alias{make_gt_grad_pal}
\title{Title}
\usage{
make_gt_grad_pal(data, ..., max_colours)
}
\arguments{
\item{data}{Your data frame or tibble object}

\item{max_colours}{}
}
\description{
Title
}
