% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gt_summary.R
\name{make_gt_summary_table}
\alias{make_gt_summary_table}
\title{Create a summary table, one step of the gt summary process}
\usage{
make_gt_summary_table(data, group_var, sentiment_var)
}
\arguments{
\item{sentiment_var}{}
}
\description{
Create a summary table, one step of the gt summary process
}
