% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_top_terms_table.R
\name{make_top_terms_table}
\alias{make_top_terms_table}
\title{Quickly make a top terms table from UMAP/Sentence Transformers/kMeans clusters}
\usage{
make_top_terms_table(df, group_var = cluster_name, text_var = mention_content)
}
\arguments{
\item{df}{Data frame object with clusters/topics}

\item{group_var}{cluster/topic variable}

\item{text_var}{the text variable}
}
\value{
a data frame with tokens counted
}
\description{
Quickly make a top terms table from UMAP/Sentence Transformers/kMeans clusters
}
