% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{net_income}
\alias{net_income}
\title{Rough estimate of take-home salary for different patypoints with student loan and pension taken into account.}
\usage{
net_income(
  salary,
  pension = TRUE,
  student_loan = TRUE,
  tax_free_allowance = 12570
)
}
\arguments{
\item{salary}{Gross salary}

\item{pension}{Contributing to a pension or not? If TRUE then value is 0.05 * taxable income(may edit this to be a prop.)}

\item{student_loan}{Repaying a student loan or not? If TRUE then value is 0.05 (may edit this to be a prop.)}

\item{tax_free_allowance}{}
}
\value{
integer with net pay per year (/12 for monhtly)
}
\description{
Rough estimate of take-home salary for different patypoints with student loan and pension taken into account.
}
