% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_par_fit_LDAs.R
\name{old_par_fit_LDAs}
\alias{old_par_fit_LDAs}
\title{Parallelised fit_LDAs function + coherence included.}
\usage{
old_par_fit_LDAs(dtms, k_opts = 2:3, iter_opts = 2000, coherence_n = 10)
}
\arguments{
\item{dtms}{A nested tibble where each row is a dtm}

\item{k_opts}{vector of the different values of k to fit an LDA model with (k = number of topics)}

\item{iter_opts}{NA vector of the different values of iter to fit an LDA model with}

\item{coherence_n}{Number of words to calculate coherence from}
}
\value{
A nested tibble in which each row contains an LDA model.
}
\description{
Fits multiple LDA models (one for each combination of the supplied parameters)
This is generally faster than fit_LDAs, the larger the dataset, the larger the benefit.
}
