% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_summ.R
\name{percent_summ}
\alias{percent_summ}
\title{Count a factor and then add a column which shows percentage}
\usage{
percent_summ(df, var)
}
\arguments{
\item{df}{Data frame or tibble object}

\item{var}{The variable to count}
}
\value{
A summary table with the factor, count and percentage
}
\description{
Count a factor and then add a column which shows percentage
}
