% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bars_coherence.R
\name{plot_bars_coherence}
\alias{plot_bars_coherence}
\title{Plot intra-topic coherence bar charts}
\usage{
plot_bars_coherence(lda, selected_k, ...)
}
\arguments{
\item{lda}{Latent Dirichlet Allocation object}

\item{selected_k}{User-determined value of k (# of topics)}

\item{...}{Additional arguments which are passed down at every level}
}
\value{
ggplot2 object
}
\description{
Plot intra-topic coherence bar charts
}
