% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_bigrams.R
\name{plot_group_bigrams}
\alias{plot_group_bigrams}
\title{Make a bigram network for each level of a categorical variable}
\usage{
plot_group_bigrams(data, cat_var, text_var = mention_content, ...)
}
\arguments{
\item{data}{Data frame or tibble object}

\item{cat_var}{The categorical variable to create plots for}

\item{text_var}{Name of the text variable}

\item{...}{Additional arguments for make_bigram_viz i.e. n = , top_n =, min =}
}
\value{
list of ggplot2 objects
}
\description{
Make a bigram network for each level of a categorical variable
}
