% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sent_percent.R
\name{plot_group_sentiment}
\alias{plot_group_sentiment}
\title{Quickly plot group sentiment distributions as a perentage}
\usage{
plot_group_sentiment(df, group_var = topic, sentiment_var = sentiment)
}
\arguments{
\item{df}{data frame}

\item{group_var}{Grouping variable, e.g. country, topic, cluster}

\item{sentiment_var}{Sentiment variable (categorical)}
}
\value{
Ggplot stacked bar chart with x and y coords flipped
}
\description{
Quickly plot group sentiment distributions as a perentage
}
