% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_vol_time.R
\name{plot_group_vol_time}
\alias{plot_group_vol_time}
\title{Quickly plot faceted volume of groups over time}
\usage{
plot_group_vol_time(
  df,
  group_var = group,
  date_var = date,
  unit = "week",
  nrow = 2
)
}
\arguments{
\item{df}{Data frame or tibble}

\item{group_var}{grouping variable e.g. country, cluster, topic etc.}

\item{date_var}{Variable which contains date information (can be datetime too I think)}

\item{unit}{Time unit fed into lubridate::floor_date e.g. "week", "day", "month", "second", "quarter", "year"}

\item{nrow}{How many rows the plot should be shown in}
}
\value{
ggplot object of facetted bar charts
}
\description{
Quickly plot faceted volume of groups over time
}
