% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_volume_lollipop.R
\name{plot_group_volume_lollipop}
\alias{plot_group_volume_lollipop}
\title{Quick plot for a topic (or any group) volume lollipop chart}
\usage{
plot_group_volume_lollipop(df, group_var = name, project_name = "project")
}
\arguments{
\item{df}{your data frame or tibble object}

\item{group_var}{the variable you wish to produce grouped volume for, e.g. topic, or name (if pivot_longer and not renaming)}

\item{project_name}{a string which will be pasted into the title, e.g. "project_name = Internal Project" gives "Internal Project - Topic Volume" as title.}
}
\value{
a ggplot object
}
\description{
Quick plot for a topic (or any group) volume lollipop chart
}
\examples{
\dontrun{
Example 1:
plot_group_volume_lollipop(df, group_var = topic, project_name = "My Private Project")

Example 2:
plot_group_volume_lollipop(df, group_var = social_network,
project_name = "Sportswear Landscape Analysis")
}
}
