% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_wlos_network.R
\name{plot_group_wlos_network}
\alias{plot_group_wlos_network}
\title{Network plot of a grouping variable's Top Terms by WLOS}
\usage{
plot_group_wlos_network(
  data,
  group_var = brand,
  text_var = message,
  n_terms = 20,
  with_ties = FALSE
)
}
\arguments{
\item{data}{Data Frame or Tibble object}

\item{group_var}{The variable you wish to group by - e.g. sentimnet, brand, entity}

\item{text_var}{The text variable}

\item{n_terms}{Number of terms to allow in plot}

\item{with_ties}{Whether to allow for ties or not (possibly selecting more than \code{n_terms})}
}
\value{
A network visualisation
}
\description{
Similar to plot_group_terms_network but filtered by WLOs rather than by frequency.
}
\examples{
\dontrun{
Example:
df \%>\%
    plot_group_wlos_network(
    group_var = sentiment,
    text_var = mention_content,
    n_terms = 20,
    with_ties = FALSE)
}
}
