% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_k_coherence.R
\name{plot_k_coherence}
\alias{plot_k_coherence}
\title{Scree plot for topic coherence analysis}
\usage{
plot_k_coherence(lda)
}
\arguments{
\item{lda}{Latent Dirichlet Allocation object}
}
\value{
ggplot2 object
}
\description{
Scree plot for topic coherence analysis
}
