% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_save.R
\name{plot_save}
\alias{plot_save}
\title{Save plots using ggsave but with some default arguments sorted}
\usage{
plot_save(
  plot = ggplot2::last_plot(),
  dir,
  height = 6,
  width = 8,
  dpi = 320,
  units = c("in", "cm", "mm", "px")
)
}
\arguments{
\item{plot}{Which plot to save? Either an object or default to the \code{last_plot()}}

\item{dir}{What should the plot be saved as, and where?}

\item{height}{Plot's height (default in inches)}

\item{width}{Plot's width (default in inches)}

\item{dpi}{Resolution}

\item{units}{Measurement unit e.g. "in" = inches}
}
\value{
Saves a plot in the appropriate directory
}
\description{
Save plots using ggsave but with some default arguments sorted
}
\examples{
\dontrun{
mtcars \%>\%
  ggplot(aes(x= cyl, y = mpg)) +
  geom_point()

plot_save("viz/imaginary_plot.png", dpi = 700)
}
}
