% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_terms.R
\name{plot_top_terms}
\alias{plot_top_terms}
\title{Make faceted lollipop charts for top terms from a cluster}
\usage{
plot_top_terms(
  top_terms_table,
  words_var = words,
  group_var = cluster_name,
  top_n = 15,
  nrow = 2
)
}
\arguments{
\item{top_terms_table}{the output of make_top_terms_table()}

\item{words_var}{the name of the variable in which word tokens have been saved}

\item{group_var}{cluster / topic / grouping variable}

\item{top_n}{number of terms per plot}

\item{nrow}{number of rows to display plots on}
}
\value{
faceted lollipops
}
\description{
Make faceted lollipop charts for top terms from a cluster
}
