% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pull.R
\name{sample_pull}
\alias{sample_pull}
\title{Quickly view a sample of any variable}
\usage{
sample_pull(df = df, variable = mention_content, n = 10)
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{variable}{The variable you want to sample}

\item{n}{Number of samples you want to print}
}
\value{
Printed samples of your variable
}
\description{
Quickly view a sample of any variable
}
