% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics_rename.R
\name{topics_rename}
\alias{topics_rename}
\title{Rename a list of topics to a new list and set the variable as a factor with new names as levels}
\usage{
topics_rename(df, topic_var, old_names, new_names)
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{topic_var}{The current variable the topics are stored in}

\item{old_names}{The old names of the topics}

\item{new_names}{New names of the topics}
}
\value{
Data Frame or Tibble with topic variable changed in place
}
\description{
Rename a list of topics to a new list and set the variable as a factor with new names as levels
}
\examples{
\dontrun{
Example 1:
old_names <- paste0("topic_", 1:2)
new_names <- c("New Topic 1", "New Topic 2")
df \%>\% topics_rename(topic_var = name, old_names = old_names, new_names = new_names)

}
}
