% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_shiny.R
\name{umap_shiny}
\alias{umap_shiny}
\title{Builds a Shiny app that can be used to label embedded texts quickly}
\usage{
umap_shiny(data, text_var, colour_var, ...)
}
\arguments{
\item{data}{Your data frame, must have a V1 and V2 variable (2d UMAP output)}

\item{text_var}{Text variable}

\item{colour_var}{Variable colur should be mapped to}

\item{...}{currently unused, may go to select ahead of \code{dataTableOutput()}}
}
\value{
shiny app
}
\description{
Builds a Shiny app that can be used to label embedded texts quickly
}
