% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gt_summary.R
\name{disp_gt_summary}
\alias{disp_gt_summary}
\title{Create a gt_summary table fit for use in client briefs}
\usage{
disp_gt_summary(data, sentiment_var, group_var, max_colours)
}
\arguments{
\item{max_colours}{}
}
\description{
Function calls various helper functions, @seealso \code{make_gt_gradient_palette},
\code{make_gt_sumamry_table2}
}
