% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_fit_LDAs.R
\name{par_fit_LDAs}
\alias{par_fit_LDAs}
\title{Parallelised fit_LDAs function + coherence included}
\usage{
par_fit_LDAs(dtms, k_opts = 2:3, iter_opts = 2000, coherence_n = 10)
}
\arguments{
\item{dtms}{Document Term Matrix=}

\item{k_opts}{Number of topics to generate}

\item{iter_opts}{Number of iterations for Gibbs sampler}

\item{coherence_n}{Number of words to calculate cohernece fron}
}
\value{
tibble
}
\description{
To access fast processing in parallel, load the future library and call:
plan(multisession(workers = availableCores()))
}
