% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_sentiment.R
\name{plot_group_sentiment}
\alias{plot_group_sentiment}
\title{Quickly plot group sentiment distributions as a perentage}
\usage{
plot_group_sentiment(
  df,
  group_var = topic,
  sentiment_var = sentiment,
  type = "percent",
  title = "Grouped Sentiment Chart",
  bar_labels = c("None", "Percent", "Volume")
)
}
\arguments{
\item{df}{data frame}

\item{group_var}{Grouping variable, e.g. country, topic, cluster}

\item{sentiment_var}{Sentiment variable (categorical)}

\item{type}{Whether the plot should be of volume or py percentage. Acccepts "percent" or "volume"}

\item{title}{The title of the plot, entered as a string.}

\item{bar_labels}{Whether to add the raw volume, percentage or neither to the bars}
}
\value{
Ggplot stacked bar chart with x and y coords flipped
}
\description{
Quickly plot group sentiment distributions as a perentage
}
\examples{
\dontrun{
Example 1:
plot_group_sentiment(df, group_var = platform, sentiment_var = sentiment, type = "percent")

Example 2:
plot_group_sentiment(df, group_var = topic, sentiment_var = sentiment, type = "volume")
}
}
