% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot_list.R
\name{save_plot_list}
\alias{save_plot_list}
\title{Save a list of plots}
\usage{
save_plot_list(plot_list, plot_type = "bigram")
}
\arguments{
\item{plot_list}{Should be a named list of plots}

\item{plot_type}{For saving prefix that gets pasted with name and .png = e.g. "bigram_microsoft_topic_1.png"}
}
\value{
does not return anything - saves plots to working directory
}
\description{
Save a list of plots
}
